wip(cli): new player
