+++ .pre-commit-config.yaml
@@ -27,6 +27,7 @@ repos:
   rev: v0.0.284
   hooks:
   - id: ruff
+    args: [--fix]
 - repo: https://github.com/pre-commit/mirrors-mypy
   rev: v1.5.0
   hooks:
